Hopefully resolving error to upload to GitHub.
Current error has been about not being able to
push refs to GitHub.
Researching this topic, I see that my local
branch is behind the remote one. Which translates
to: commits have been made, even though the system
doesn't remember, and not updated to local file.
So here I am trying to make an actual commit to
and send the update.